+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 - Refactor the tokenfactory wasmbinding into its own path
 - Refactor the wasmbinding implementation to allow multiple msg and query types
 - Add EVM wasmbinding queries
+- Add bech32 wasmbinding queries